- Refactor pawn moves out of rules - Started rook rules
